refactor: extract registerEndpoint function
